Fixing issues with column sizes (#1393)

+++ tracer/arithmetization/src/main/java/net/consensys/linea/zktracer/module/hub/TransactionStack.java
@@ -85,7 +85,7 @@ public class TransactionStack implements StackedContainer {
     }
   }
 
-  public int getAccumulativeGasUsedInBlockBeforeTxStart() {
+  public long getAccumulativeGasUsedInBlockBeforeTxStart() {
     return this.relativeTransactionNumber == 1 ? 0 : this.previous().getAccumulatedGasUsedInBlock();
   }
 }
